refactor: extract useServerInfo hook — App.tsx back under the 800-line cap
The merge with main (#92 added vision-tool UI lines to App.tsx) pushed
it to 801; move the self-contained /api/system/info state+effect into
hooks/useServerInfo.ts, matching useUpdateCheck/useSessionKeepalive.
